backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
psycopg2-binary==2.9.9
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.1
python-multipart==0.0.19
boto3==1.35.90
pydantic==2.9.2
pydantic[email]==2.9.2
python-dotenv==1.0.1
pgvector==0.3.5
numpy==1.26.4

# Machine Learning - Biometria Real
transformers==4.46.3
torch==2.5.1
torchvision==0.20.1
timm==1.0.11
Pillow==10.4.0
opencv-python==4.10.0.84

